requirements.txt
click==8.0.1
dataclasses==0.6
Flask==2.0.1
Flask-MySQLdb==0.2.0
gunicorn==20.1.0
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
